Support bundling all transitive crates in one directory on Windows
Transitive crate dependencies are passed to rustc by specifying the folder they
live in as a library search path (an `-Ldependency=path` flag). Given that
bazel builds all libraries into their own folders, this means passing
potentially hundreds of paths to rustc.

On Windows, when the number of transitive dependencies is too great, the PATH
environment variable overflows its limit of 32k characters and rustc runs into
`LoadLibraryExW` errors, failing the build.

For this reason, this patch can optionally create symlinks to all the
transitive crates into a single directory that it then sets as the
`-Ldependency=` flag to rustc. We only do this when building on Windows and
when the sum of the path lengths exceeds 16k, which happens for a handful of
top-level crates in the dropbox_core project.